create controllers and fix bugs

+++ src/main/java/com/project/ecommerce/entity/Item.java
@@ -18,7 +18,7 @@ public class Item implements Serializable {
     private String description;
 
     @Column(name = "price")
-    private Integer price;
+    private Float price;
 
     public Long getId() {
         return id;
@@ -32,7 +32,7 @@ public class Item implements Serializable {
         return description;
     }
 
-    public Integer getPrice() {
+    public Float getPrice() {
         return price;
     }
 }
